switch StackOnUlist to Qt
