--- StackOnUList.sln
@@ -1,20 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 11.00
-# Visual Studio 2010
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "StackOnUList", "StackOnUList\StackOnUList.vcxproj", "{4D2B4612-4FA0-46AC-85EF-193706622702}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Win32 = Debug|Win32
-		Release|Win32 = Release|Win32
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{4D2B4612-4FA0-46AC-85EF-193706622702}.Debug|Win32.ActiveCfg = Debug|Win32
-		{4D2B4612-4FA0-46AC-85EF-193706622702}.Debug|Win32.Build.0 = Debug|Win32
-		{4D2B4612-4FA0-46AC-85EF-193706622702}.Release|Win32.ActiveCfg = Release|Win32
-		{4D2B4612-4FA0-46AC-85EF-193706622702}.Release|Win32.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
