--- StackOnUList/StackOnUList/StackOnUList.pro
@@ -1,7 +1,0 @@
-TEMPLATE = app
-CONFIG += console
-CONFIG -= app_bundle
-CONFIG -= qt
-
-SOURCES += main.cpp
-
