--- StackOnUList/StackOnUList/StackOnUList_main.cpp
@@ -1,10 +1,0 @@
-#include <iostream>
-
-using namespace std;
-
-int main()
-{
-    cout << "Hello World!" << endl;
-    return 0;
-}
-
