--- StackOnUList/StackOnUList/mystack.cpp
@@ -1,5 +1,0 @@
-#include "mystack.h"
-
-MyStack::MyStack()
-{
-}
