--- StackOnUList/StackOnUList/mystack.h
@@ -1,10 +1,0 @@
-#ifndef MYSTACK_H
-#define MYSTACK_H
-
-class MyStack
-{
-public:
-    MyStack();
-};
-
-#endif // MYSTACK_H
