--- _StackOnUList/stdafx.h
@@ -1,15 +1,0 @@
-// stdafx.h : include file for standard system include files,
-// or project specific include files that are used frequently, but
-// are changed infrequently
-//
-
-#pragma once
-
-#include "targetver.h"
-
-//#include <stdio.h>
-//#include <tchar.h>
-
-
-
-// TODO: reference additional headers your program requires here
